feat: Enhance dashboard functionality with WebSocket integration and value formatting
